+++ persistence/actor_shared_sql.go
@@ -4,9 +4,6 @@ import (
 	"context"
 	"encoding/hex"
 	"fmt"
-	"github.com/pokt-network/pocket/shared/converters"
-	"math/big"
-
 	"github.com/jackc/pgx/v5"
 	"github.com/pokt-network/pocket/persistence/types"
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
@@ -17,11 +14,10 @@ import (
 // for the purpose of being explicit: https://github.com/pokt-network/pocket/pull/140#discussion_r939731342
 // TODO Cleanup with #149
 const (
-	UndefinedStakingStatus  = int32(0)
-	UnstakingStatus         = int32(1)
-	StakedStatus            = int32(2)
-	UnstakedStatus          = int32(3)
-	defaultAccountAmountStr = "0"
+	UndefinedStakingStatus = int32(0)
+	UnstakingStatus        = int32(1)
+	StakedStatus           = int32(2)
+	UnstakedStatus         = int32(3)
 )
 
 func UnstakingHeightToStatus(unstakingHeight int64) int32 {
@@ -86,19 +82,6 @@ func (p *PostgresContext) GetActorsUpdated(actorSchema types.ProtocolActorSchema
 	return
 }
 
-func (p *PostgresContext) getAccountAmount(accountSchema types.ProtocolAccountSchema, address string, height int64) (amount string, err error) {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return
-	}
-	amount = defaultAccountAmountStr
-	if err = tx.QueryRow(ctx, accountSchema.GetAccountAmountQuery(address, height)).Scan(&amount); err != pgx.ErrNoRows {
-		return
-	}
-
-	return amount, nil
-}
-
 func (p *PostgresContext) getActor(actorSchema types.ProtocolActorSchema, address []byte, height int64) (actor *coreTypes.Actor, err error) {
 	ctx, tx, err := p.getCtxAndTx()
 	if err != nil {
@@ -157,80 +140,6 @@ func (p *PostgresContext) getChainsForActor(
 		actor.Chains = append(actor.Chains, chainID)
 	}
 	return actor, nil
-}
-
-func (p *PostgresContext) operationAccountAmount(
-	accountSchema types.ProtocolAccountSchema,
-	address, amount string,
-	op func(*big.Int, *big.Int) error) error {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return err
-	}
-	height, err := p.GetHeight()
-	if err != nil {
-		return err
-	}
-	originalAmount, err := p.getAccountAmount(accountSchema, address, height)
-	if err != nil {
-		return err
-	}
-	originalAmountBig, err := converters.StringToBigInt(originalAmount)
-	if err != nil {
-		return err
-	}
-	amountBig, err := converters.StringToBigInt(amount)
-	if err != nil {
-		return err
-	}
-	if err := op(originalAmountBig, amountBig); err != nil {
-		return err
-	}
-	if _, err = tx.Exec(ctx, accountSchema.InsertAccountQuery(address, converters.BigIntToString(originalAmountBig), height)); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (p *PostgresContext) getAccountsUpdated(accountType types.ProtocolAccountSchema, height int64) (accounts []*coreTypes.Account, err error) {
-	query := accountType.GetAccountsUpdatedAtHeightQuery(height)
-
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return
-	}
-
-	rows, err := tx.Query(ctx, query)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		acc := new(coreTypes.Account)
-		if err = rows.Scan(&acc.Address, &acc.Amount); err != nil {
-			return nil, err
-		}
-		accounts = append(accounts, acc)
-	}
-
-	return
-}
-
-func (p *PostgresContext) insertAccount(accountType types.ProtocolAccountSchema, address, amount string) error {
-	ctx, tx, err := p.getCtxAndTx()
-	if err != nil {
-		return err
-	}
-	height, err := p.GetHeight()
-	if err != nil {
-		return err
-	}
-	// DISCUSS(team): Do we want to panic if `amount < 0` here?
-	if _, err = tx.Exec(ctx, accountType.InsertAccountQuery(address, amount, height)); err != nil {
-		return err
-	}
-	return nil
 }
 
 func (p *PostgresContext) InsertActor(actorSchema types.ProtocolActorSchema, actor *coreTypes.Actor) error {
